TDS-5 Add registration workflow
Added Javascript to the consent page so that you can't submit until you have checked all of the consent boxes.  Have also added code to do some checking on the contents of the page, and display error messages if there is a content problem.  Have not yet added code to create a new person/patient.